requirements.txt
fastapi
ffmpeg-python
uvicorn
flake8
sqlalchemy
databases[postgresql]
asyncpg
pydantic
onnxruntime
opencv-python
numpy
pillow
torchvision
